# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  templatev25_internal_http_dto.OrgTypeAddSystemsDto:
    properties:
      system_ids:
        items:
          type: integer
        minItems: 1
        type: array
      type_id:
        type: integer
    required:
    - system_ids
    - type_id
    type: object
  templatev25_internal_http_dto.OrgTypeRolesAddDto:
    properties:
      role_ids:
        items:
          type: integer
        type: array
      type_id:
        description: org type id
        type: integer
    required:
    - role_ids
    - type_id
    type: object
  templatev25_internal_http_dto.OrgUserCreateDto:
    properties:
      org_id:
        type: integer
      user_id:
        type: integer
    required:
    - org_id
    - user_id
    type: object
  templatev25_internal_http_dto.OrgUserDeleteDto:
    properties:
      org_id:
        type: integer
      user_id:
        type: integer
    required:
    - org_id
    - user_id
    type: object
  templatev25_internal_http_dto.OrganizationDto:
    properties:
      address_detail:
        maxLength: 255
        type: string
      aimag_id:
        type: integer
      aimag_name:
        maxLength: 255
        type: string
      bag_id:
        type: integer
      bag_name:
        maxLength: 255
        type: string
      country_code:
        type: string
      country_name:
        type: string
      country_name_en:
        type: string
      email:
        maxLength: 50
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      name:
        maxLength: 255
        type: string
      parent_address_id:
        type: integer
      parent_address_name:
        maxLength: 25
        type: string
      parent_id:
        type: integer
      phone_no:
        maxLength: 8
        type: string
      reg_no:
        maxLength: 7
        type: string
      sequence:
        type: integer
      short_name:
        maxLength: 255
        type: string
      sum_id:
        type: integer
      sum_name:
        maxLength: 255
        type: string
      type_id:
        type: integer
    required:
    - name
    - reg_no
    - type_id
    type: object
  templatev25_internal_http_dto.OrganizationTypeDto:
    properties:
      code:
        maxLength: 255
        type: string
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - code
    - name
    type: object
  templatev25_internal_http_dto.OrganizationUpdateDto:
    properties:
      address_detail:
        maxLength: 255
        type: string
      aimag_id:
        type: integer
      aimag_name:
        maxLength: 255
        type: string
      bag_id:
        type: integer
      bag_name:
        maxLength: 255
        type: string
      country_code:
        type: string
      country_name:
        type: string
      country_name_en:
        type: string
      email:
        maxLength: 50
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      name:
        maxLength: 255
        type: string
      parent_address_id:
        type: integer
      parent_address_name:
        maxLength: 25
        type: string
      parent_id:
        type: integer
      phone_no:
        maxLength: 8
        type: string
      reg_no:
        maxLength: 7
        type: string
      sequence:
        type: integer
      short_name:
        maxLength: 255
        type: string
      sum_id:
        type: integer
      sum_name:
        maxLength: 255
        type: string
      type_id:
        type: integer
    required:
    - name
    - reg_no
    - type_id
    type: object
...
info:
  contact: {}
  description: Fiber + Gorm + Auth + RBAC demo
  title: Template Backend API
  version: "1.0"
paths:
...
  /organization:
    get:
      description: Get paginated list of organizations
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List organizations
      tags:
      - organization
    post:
      consumes:
      - application/json
      description: Create a new organization
      parameters:
      - description: Organization data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrganizationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - organization
  /organization/{id}:
    delete:
      description: Delete an organization (soft delete)
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - organization
    put:
      consumes:
      - application/json
      description: Update an existing organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrganizationUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - organization
  /organization/find:
    get:
      description: Search organization from SSO Core service
      parameters:
      - description: Search text (reg_no or name)
        in: query
        name: search_text
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Find organization from Core
      tags:
      - organization
  /organization/tree:
    get:
      description: Get hierarchical organization tree
      parameters:
      - description: Organization ID
        in: query
        name: org_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get organization tree
      tags:
      - organization
  /orgtype:
    get:
      description: Get paginated list of organization types
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List organization types
      tags:
      - orgtype
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrganizationTypeDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create organization type
      tags:
      - orgtype
  /orgtype/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete organization type
      tags:
      - orgtype
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrganizationTypeDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update organization type
      tags:
      - orgtype
  /orgtype/role:
    get:
      parameters:
      - description: Organization Type ID
        in: query
        name: type_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get roles by organization type
      tags:
      - orgtype
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrgTypeRolesAddDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add roles to organization type
      tags:
      - orgtype
  /orgtype/system:
    get:
      parameters:
      - description: Organization Type ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get systems by organization type
      tags:
      - orgtype
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrgTypeAddSystemsDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add systems to organization type
      tags:
      - orgtype
  /orguser:
    delete:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrgUserDeleteDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove user from organization
      tags:
      - orguser
    get:
      parameters:
      - description: Organization ID
        in: query
        name: org_id
        type: integer
      - description: User ID
        in: query
        name: user_id
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List organization users
      tags:
      - orguser
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.OrgUserCreateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add user to organization
      tags:
      - orguser
  /orguser/organizations:
    get:
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: integer
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get organizations by user
      tags:
      - orguser
  /orguser/users:
    get:
      parameters:
      - description: Organization ID
        in: query
        name: org_id
        type: integer
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get users by organization
      tags:
      - orguser
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"